Clear the dompurify advisory, ignore unfixable image-size, restore the LOC gate
osv-scanner flagged dompurify 3.4.12 (patched in 3.4.13) plus two image-size
denial-of-service advisories. Bump the dompurify pin and override in the root
and ui manifests, and update the pin assertions that name the expected
release. image-size reaches apps/mobile only through metro, the Expo bundler
that runs at build time and never parses untrusted images; every published
version through 2.0.2 is affected with no patched release, so both advisories
get dated ignores alongside the existing unfixable transitives.

Wiring background skill capture had pushed chat-runtime.ts back over the
2000-line gate. Move the three pending-chat-completion helpers to
chat-runtime-state.ts, which already owns the pendingChatCompletions map they
are the only consumers of.
