+++ package.json
@@ -134,7 +134,7 @@
     "basic-ftp": "5.3.1",
     "bn.js": "5.2.4",
     "brace-expansion": "1.1.18",
-    "dompurify": "3.4.12",
+    "dompurify": "3.4.13",
     "flatted": "3.4.2",
     "follow-redirects": "1.16.0",
     "form-data": "4.0.6",
