+++ ui/package.json
@@ -54,7 +54,7 @@
   },
   "overrides": {
     "brace-expansion": "5.0.9",
-    "dompurify": "3.4.12",
+    "dompurify": "3.4.13",
     "postcss": "8.5.23"
   }
 }
